requirements.txt
transformers==4.20
datasets==2.5.0
tqdm==4.64.0
pytorch_lightning==1.9.4
ninja==1.10.2
spacy==3.5.1
apache_beam==2.46.0
wandb==0.14.0
git+https://github.com/openai/CLIP.git#egg=clip
h5py==3.8.0
timm==0.6.13
sentencepiece==0.1.98

einops==0.6.0
einops_exts==0.0.4
resize_right==0.0.2
PyWavelets==1.4.1
pytorch_wavelets==1.3.0
